+++ src/components/Popover/index.ts
@@ -1,5 +1,4 @@
-// Popover from copy-ui @ 2025-10-28.
+// Popover from copy-ui @ 2025-11-15.
 
-export type { Placement } from "./popover";
-export { Popover } from "./popover";
+export { type Placement, Popover } from "./popover";
 export type { PopoverTriggerRender } from "./popover-trigger";