+++ src/components/Popover/CHANGELOG.md
@@ -14,3 +14,5 @@
 - 2025-10-27: Make `Popover.Trigger`'s render can have prop `isOpen`.
 - 2025-10-28: Using `Modal.Raw` rather than `Modal`. Export the type
   `PopoverTriggerRender`.
+- 2025-11-15: Add `offset` prop to `Popover` to control the distance between the
+  trigger and the portal.